+++ packages/admin/src/pages/scheme/[schemeId]/editors.getServerSideProps.tsx
@@ -64,11 +64,11 @@ const getEditorRows = async ({
 const getEditorsServerSideProps = async ({
   params,
   req,
+  query,
 }: GetServerSidePropsContext) => {
   const { schemeId } = params as Record<string, string>;
   const sessionCookie = getSessionIdFromCookies(req);
   const userServiceJwt = getUserTokenFromCookies(req);
-  const queryParams = new URLSearchParams(req.url?.split('?')[1] ?? '');
 
   const isOwner = await isSchemeOwner(schemeId, sessionCookie);
 
@@ -92,7 +92,7 @@ const getEditorsServerSideProps = async ({
         sessionCookie,
         userServiceJwt,
       }),
-      newEditor: queryParams.get('newEditor'),
+      newEditor: query['newEditor'],
     },
   };
 };
