+++ packages/admin/src/services/SchemeEditorService.ts
@@ -31,3 +31,17 @@ export const getSchemeEditors = async (
   );
   return data;
 };
+
+export const addSchemeEditor = async (
+  schemeId: string,
+  sessionId: string,
+  userToken: string,
+  editorEmailAddress: string
+) => {
+  const response = await axios.post(
+    `${SCHEME_HOST}/${schemeId}/editors`,
+    { editorEmailAddress },
+    getFullConfig(sessionId, userToken)
+  );
+  return response.data;
+};
